test_bed: Roll back some moves into cc65 libraries.
I'm starting to realize that I probably don't want to enable the full
functionality provided by the cc65 libraries, as it seems to want to
control a bit too much for my tastes.  It seems to be more targeted
towards running like an embedded program, as opposed to just providing
a C compiler for you to do whatever with.

As of right now, the only piece I can identify that we really can't live
without from the cc65 libraries is the setup of their special "sp" stack
pointer.  This is an in-memory variable, not the regular stack pointer
register, and C programs built with cc65 apparently _need_ for it to be
set up properly.

I think we can live without the constructor/destructor stuff, but I've
had to leave it intact for now to get things to build.  The
runtime/condes.s file from the cc65 libraries expects the linker-defined
symbols relevant to the constructor/destructor stuff to exist, even if
there aren't any actual constructors or destructors.  I'm sure there's a
way to get rid of that stuff completely, but it's still here for now.