Start working on the skeleton of my port
I've done a lot of work to get the existing Uzi code building, but I'm
realizing that I have a lot of foundation to lay before I can really
start trying to run that code.  I intend to re-use as much of it as
makes sense, but for now I want to shift focus and start working on the
actual NES side of things, to get a basic C program working, and then
some kind of putc/printk implementation written.

This is the first step, to clone the working hello_mmc5 prototype,
change some names, and add some Makefile rules to build C files.  Some
of that is untested, but the .nes file does still build and run, so
that much is still working.