WIP putc/buffer dump working once
This small change made a huge difference, maybe for obvious reasons.
I'm now seeing the HELLO string print, and it appears to be dumping some
code or other random bytes as well after a few passes.  Eventually the
pallette gets stomped over and the screen goes white, but it's pretty
obvious that this is more fallout from PPU dump loop running wild.