Fix screenbuffer copy
There were a few problems here.  First off, not casting &screenbuf as a
char * when doing arithmetic on it apparently makes it treat screenbuf
as a 2-byte value, so screenbuf + 10 ends up doing screenbuf + (10 * 2),
which is obviously not what we want.  I'm still not sure I fully
understand this, but this change does fix the issue.

The second problem is that we start writing the to PPU memory at 0x2020
to skip the first 32 tiles (which we can't see), but we still reserve
and copy a full 960 bytes for our "external" screenbuffer.  This ends up
stomping past the end of the nametable into the attribute table, causing
some tiles in the upper left to get screwed up after the first full
buffer dump.  Fix this by trimming down the screenbuffer to 896 bytes.
That's 960 bytes minus 64 bytes for the two non-visible rows of tiles.