kernel.c: Fix putc()
This function wasn't actually writing to the screenbuf array.  I changed
the following line:

	screenbuf[cursor_pos] = ptable_c;

To:

	*((char *) &screenbuf + cursor_pos) = ptable_c;

As a shot-in-the-dark attempt to fix the problem.  Surprisingly, this
worked.  After staring at it for a while, I decided to try explicitly
declaring screenbuf as an array instead of a char *.  This seems to do
the trick, and gets the array assignment in putc() working properly.  I
think this would have worked fine with it declared as a char * in gcc,
so just a bit of a learning experience here...