font: Create my own font and modify the code to use it.
I've also checked in a script that I had to use to correct a mistake I
made in nestile.  I accidentally skipped the ';' character, so
everything after that was off by one.  nestile didn't have a way to
copy/move tiles so I just wrote that C program to shift all the tiles.

It's purpose built for that task, but I felt like I might get use out of
the code later.