lib/string: Fix strlen() implementation
I'm not entirely positive what I got wrong with the order of operations
on my pointer arithmetic, but it was obviously something.  This change
gets strlen() to return the correct length in all cases, or at least the
ones I've seen.  This, in turn, fixes other weird bugs beneath kprintf.