Make sure that header changes for a C file rebuild
This isn't the ideal way to do this, but it does work.  Note that it
will trigger unnecessary rebuilds, as every C file will be rebuilt
whenever any header is changed.  This is better than the weird issues
you'll get if C files aren't rebuilt after header changes, but needs to
be addressed at some point.