Copy in doexit/chksigs/wakeup/addtick
Also stub in i_deref and doclose.

This is getting stuck in and infinite loop in getproc, because tty_read
will never read anything from the UART, since it doesn't exist.  From
what I can tell, the UART would normally contain a "boot character"
which would get written into bootchar, and then used as the device
number for ROOTDEV.

This obviously can't work as expected in our environment, so I'll hack
around it.  Just noting that it's hosing the system right now.