devtty: Insert a dummy character to initialize the queue.
The cdread() call in init2() was dropping into an infinite psleep()
because there was nothing in the ttyinq to read.  Create a tty_init()
function to insert a dummy character into the queue, just to give us
something to pull out.  This ultimately needs to be rethought a bit, but
for not this gets things working as expected.