head.S: Clear vblank bit in PPU_STATUS before re-enabling NMI
If the vblank bit is set in PPU_STATUS when the vblank NMI is enabled,
an NMI will be generated immediately.  If we're too late in vblank to
finish PPU operations during the NMI, we can see subtle graphical errors
because of this.  Avoid this problem by reading PPU_STATUS before
re-enabling the vblank NMI.

This issue was discovered while working on the IRQ stack switch
problems, where I disabled vblank NMI during the IRQ handler.  I'm not
positive if it should be this way, but I observed an NMI ocurring in the
middle of an IRQ when I was debugging stuff, and it definitely caused
problems.