interrupts: Implement IRQ stack switching
This is actually necessary in order to not clobber the kernel/process
stack during an IRQ, and always has been, but I wasn't aware of any
issues until the kernel got complex enough.  This is very rough code,
but it does seem to work.  I need to de-hard-code some stuff here still,
but I'm checkpointing this now because this was very tricky to get
working.