Keyboard "driver" rough draft
This is successfully reading the keyboard, but the keyboard map that I
wrote isn't quite right.  It's apparently off by 2, and some symbols are
mixed up or in the wrong place.  The core logic is working OK though, it
prints _something_ when you type, and it's responding quickly enough
that you'd be able to actually type and have it look natural.

Current output is just a spew of whatever keys were held down when the
read was done, but it shows that it works.  Needs more logic to
understand when a key was held vs. pressed.