kb: Move stuff out of interrupts.S
This obviously didn't belong in interrupts.S, it was just simpler to
work on it in there.  I've also created a subdirectory for the keyboard
stuff in the drivers folder.  My Makefile solution for this is
rudimentary, but works for now.  Eventually I should think about some
kind of more generic recursive make setup, but it's not worth the
complexity for now.