head.S: Enable NVRAM and move kernel out of it
We're currently operating in page 1 of volatile RAM (despite the comment
saying we're in page 0).  I want to use battery-backed RAM as an R/W
filesystem, so enable that, and bank switch to the first
non-battery-backed page before loading the kernel data into memory.

This is way-preliminary work for supporting multiple processes, and some
kind of useable filesystem.