POC for writes to CHR-RAM
Since we already burn most of VBlank on the screen buffer dump, we can't
write much more without stepping outside VBlank, but this is a valid
POC.  You can write a lot more if you decrease the portion of the screen
buffer that's written each VBlank.  I'll need to find a good balance
here once I actually start using this.