interrupts: Move IRQ stack down 2 pages
I had been putting the IRQ stack two pages below the normal stack, but
it doesn't need the entire range from $0100-$0600.  Move the beginning
of the IRQ stack down to $0400 to give the "normal" stack a full 1K to
grow.

I haven't seen a need for this yet, but I noticed it while working on
some other things and figured I'd go ahead and do it before I run into
problems later.