uzi-nes-utils: Fix buffer/block zeroing for superblock and dirbuf
These buffers weren't properly zeroed, and the bcopy()s for the
superblock/freelist were copying garbage off the end of their designated
buffers.  This was causing differences between the results of mkfs and a
run of fsck on the freshly made fs.  With these changes, I was able to
get my WIP fsck code to produce identical results to a fresh mkfs, which
is the desired result here.