linker: Fix alignment for PPU data buffer stuff
I really need to align the data buffer to $100 in order to be able to
efficiently process the buffer with the register space available (i.e. I
really want it to be exactly 2 pages).  I originally had the wrong
alignment in here, so fix that, and introduce the proper 'align =' bit
to the linker config to get rid of this error:

<snip>
ld65: Warning: scripts/linker/uzi.cfg:16: Segment 'DATA' isn't aligned properly; the resulting executable might not be functional.
</snip>