ppu/ppu_databuf.c: Use address-of operator for ppu_databuf pointer
I feel like this is actually wrong from a C point of view, but maybe I'm
being stupid.  Without this, we don't actually get copies into the data
buffer as expected.  I really need to try to understand what's causing
this, but this is working for now.  I'll look at this more soon, once I
figure out that other weird issues I'm having with the buffer...