ppu_databuf: Fix address-of stupidity
I realized after a while what was wrong here - this really needs to be
declared as an array, not a pointer, otherwise the math I was trying to
do is wrong.  Adding the address-of sort of "helped," but it had other
problems.

This was causing the behavior where the data buffer position would
increment twice as far as I was expecting.  This change fixes that, but
there are still other problems :sigh:

This probably needs to be fixed in other places as well...