ppu/asm/ppu_databuf.S: Fix buffer paging loop logic
My outer loop logic for write_ppubuf was wrong because iny does not set
the carry flag when it overflows.  It should, however, set the zero flag
when it wraps around, so check for this instead.

This gets PPU output way closer to working.  I can see the text that I
expect, but something is still wrong, as the graphics get hosed shortly
after boot.