ppu/ppu_databuf.c: Cut back the number of bytes to stuff
The loop currently blows past VBlank if we write more than 16
single-byte-worst-case-scenario bytes.  It's obviously not
representative of what's possible to restrict the buffer write to 16
bytes per VBlank, but it's what's necessary to get it working without
trashing the graphics for now.

I will make this smarter at some point, but for now, I think this gets
my POC for the PPU data buffer stuff working.