head.S: Actually set up nametable mirroring
I thought the header bit controlled this, but realized when learning
about scrolling that it's mapper-controller on MMC5.  The header was
already correct (although the comment was wrong), so I just need to set
the nametable mapping register ($5105) correctly.

Note that we really want vertical mirroring here, not horizontal, as
this gives the proper behavior for text scrolling, i.e. a contiguous
2000-27ff that wraps around when you scroll out past the end.