ppu/asm/ppu.S: Add CPU cycle comment to ppu_reset
This function probably wastes some cycles.  I think it currently burns
38.  I'm pretty sure this could be trimmed down, as I don't think it's
necessary to reset/write to any of these registers every VBlank unless
something has actually changed.  At the same time, it might be nearly as
expensive to check if it needs to be done as opposed to just assuming it
needs to be done every time.