ppu: Track PPU_CTRL value for consistent updates
Until now, I've only really cared about the highest bit of PPU_CTRL for
enabling/disabling VBlank, but this needs to change in order to get
scrolling working.  Create a ppu_ctrl_val symbol to track the current
value of the register, and use that when making updates to it.  This is
necessary because reads from PPU_CTRL don't actually return its current
state, from what I've observed.