ppu/asm/ppu.S: Polish scroll behavior
I'm not positive this is all completely necessary, but I want to get
this all perfect so that I know I'm not causing any bugs.

The buffered PPU_CTRL stuff is future-proofing that I know I'll need.
The PPU_STATUS read is just an extra-cautious move, as I think the
PPU_CTRL write will reset the w flag.

The scrolling changes are to fix a _very_ subtle 1-pixel gap at the top
of the screen in FCEUX.  It seems like a legit issue, as I can still see
the bottom of the previous row of characters if I try to scroll by 8
bytes (i.e. one row) instead of 9.  I don't fully understand what is
causing this.  I looked at it for a long time, and still don't get it,
so I'm going to paper over it for now...