ppu/ppu_databuf.c: Force NULL descriptors to be printed in their own frame
I had this second wait_frame in the code when I first wrote this, but I
took it out because I thought it wasn't strictly necessary.  It turns
out that, without it, there is a subtle bug when you're typing really
fast and smashing enter - if you manage to get a NULL desc queued up and
then type a couple keys after that in the same frame, the NULL write
will spill over into the next line, leaving some blanks before the next
character that gets printed.

To avoid this, force a frame wait both before and after queueing the
NULL descriptor, which, in turn, forces the NULL descriptor to be
processed in its own VBlank.  This is obviously inefficient, as there is
more than enough room for some more writes along with the single-line
blank descriptor, and there are better ways to blank the buffer during
wraparound than burning 2 frames for every line we scroll.  However,
this works fine, and gets things moving for now.