ppu/ppu_databuf.c: Move data buffer size control to queue_descriptor()
I want the queue_descriptor method to handle all of the logic to limit
what gets put in the data buffer.  For now this just means making sure
we don't write past the end of the buffer, but it will eventually have
more logic to limit the total number of descriptors, bytes written/read,
etc.