ppu_databuf: Separate data from descriptor
If I want to implement multi-byte writes to the PPU, the data shouldn't
really be baked into the descriptor, and for NULL descriptors, or
potential reads from the PPU, I don't really need to provide any data
with the descriptor, for my PoC idea anyways.

Note that I think this also fixes a subtle bug with NULL descriptors, as
there was no additional 'iny' to skip past the NULL data byte at the end
of these descriptors, which means that a NULL descriptor would _always_
terminate ppu_writebuf after it was processed.

This would never be seen with the code the way it is, since we force
NULL descriptors to get processed in their own VBlank, but it _is_
technically a bug, and could have bitten me at some point down the road.

I've also fixed up strlen() to handle NULL pointers being passed to it,
which is what will happen when we try to process a NULL descriptor (i.e.
we'll try to do strlen(NULL) when checking if there's room in the
buffer).