kb: Move dump_keyboard() functionality into clock interrupt
In order to implement something that kind of works like a real UART for
the keyboard, we need to both read the keyboard and handle the pressed
keys in interrupt context.

I looked into trying to use BRK to trigger a separate interrupt to
handle the keyboard stuff, but calling BRK from interrupt context
appears to be kind of a stupid thing to do.  It works, but it
immediately triggers the second-level interrupt, which doesn't really
buy us anything.

I also considered using the MMC5A cycle timer to trigger a different
interrupt later, outside of regular IRQ context, which might be worth
revisiting in the future, but this works for now, and doesn't require
figuring out how additional funky hardware works.