kb: Fix bugs with new KB handling
I'm realizing that the bits for the keyboard reads are inverted, so
checking the zero flag after the `ora kb_tmp` is almost always going to
cause us to behave as if a key was pressed (unless we press every key in
a single row).  I observed this by checking the return value from
read_keyboard and seeing that it was always 1.

Because of this, everything still appears to work just fine, since we're
running dump_keyboard() constantly, regardless of whether it's during an
interrupt, or in the endless loop that the kernel currently sits in.
This obviously isn't what we want though.

To fix this, XOR each row with 0xFF so that a set bit indicates a
pressed key.  This ultimately makes more sense, and simplifies the logic
to differentiate fresh keypresses from held keys a bit.

While we're at it, fix a small bug where A isn't always 0 at the
read_keyboard_out label.  Explicitly load #0 into X instead of assuming
we can achieve this with a TAX.