kb: Properly handle repeated key presses in interrupts
We need to clear prev_kb_rows every time there's a clock interrupt with
no keys pressed.  Without this, you actually can't type the same key
twice in a row, because it will always treat the key as if it has been
held across the frame (or now clock interrupt).

This used to work because we were always running dump_keyboard(),
regardless of whether or not a key was pressed, which would silently
reset the prev_kb_rows every time it was called when no keys are
being held.