interrupts: Allow NMI during IRQ
This is part of an effort to clean up my interrupt handling stuff.  Up
until now, I've just been hand-wavily enabling/disabling interrupts
where I thought it made sense, but this has created a lot of subtle
bugs, and is just ugly to look at an understand.

This is the first step in cleaning up some of the IRQ stuff.  Subsequent
commits will fix other problems that came from this change, or were
unfixable without this change.