ppu: Make databuf locking more intelligent
This turns the lock bit into a more proper spinlock, and adds some logic
to detect the case where we've tried to take the lock during an
interrupt.  We're not allowed to do this, as it's possible that normal
context has the lock, so taking it in an interrupt handler will cause a
deadlock.