ppu_databuf: Tighten up ppu locking
It turns out that we don't need to lock the PPU for nearly as long as I
was.  Pull the locking in to just around the spots where we are actually
modifying PPU state data.  This gets the PPU behaving correctly again,
but there are still some artifacts printing on the screen that I haven't
figured out yet.  I have a feeling that these are unrelated to the PPU
stuff that I've been working on.