kernel: Check for stack overflow during IRQ switch
If the normal context stack pointer has made its way up past where the
IRQ stack pointer would start (i.e. 0x400), then we're in trouble, as
we'll be scribbling over normal context stack data while handling
interrupts.

Detect this condition and halt in an infinite loop when it happens.
This isn't the ideal way to detect this, but it's easy enough to see if
this is where we're stuck in a debugger, so this is good enough for now.