kernel/process.c: Disable keyboard for now
Now that I'm actually trying to think about doing "interrupt stuff"
correctly, it's problematic to try to echo pressed keys from the
keyboard in interrupt context, since we need to take the ppu_lock as
part of ppu_putc.  If process context already has this lock, then we're
dead in the water.

My current idea for how to fix this is to implement some kind of buffer
to stuff with characters from the pressed keys, and then have the buffer
get printed at the end of the clock interrupt routine, but this is
stupid and I can fix this now.