ppu/ppu.c: Re-work ppu_lock logic
Having this be a spinlock doesn't really make sense.  For process
context, the lock should never be held, because the only other possible
holder would be interrupt context.  It seems safe to allow interrupt
context to take the lock, as long as the lock has been checked first (to
avoid a deadlock), because we have control of when we relinquish control
back to process context, i.e. if it's not taken, we know it won't get
taken while we're using it.