ppu/ppu_databuf.c: Implement cycle-based PPU data buffer throttling
Limiting the contents of the buffer by size isn't quite good enough, as
certain descriptors may fit in the buffer, but take way too long to
process.  I've hand-calculated the number of cycles needed for each
descriptor type, as well as the loop overhead that goes into processing
any descriptor, to allow us to more surgically determine what isn't safe
to load into the data buffer without a wait_frame().

I was hoping to use this to allow us to write to the buffer during
interrupts, but that's more difficult than I initially realized.  I
can't do that as things stand right now, because I need to know whether
_every_ piece of a ppu_putc will fit in the current buffer.  This might
include a newline, which runs us into problems, since we can't do part
of a ppu_putc and then try to do half of it later after getting -EAGAIN.
It has to be all or nothing, which we can't quite pull off right now.

This probably won't make sense when I read it later, but just know that
it's quite a bit more complicated than I thought to implement a "no
wait" ppu_putc.  I think we might just have to not handle keyboard input
until we can create a kernel thread to handle keyboard input in process
context instead of in interrupt context.