ppu/ppu_databuf.c: Get rid of unnecessary explicit wait_frame()s
These used to be needed to ensure that read/NULL descriptors were fully
processed before trying to do anything else, but now that we have code
to automatically force waits when the buffer is full, we don't have to
be so careful about this.

Note that we do still need an explicit wait_frame() after a read
descriptor, otherwise we don't know if the data has actually been read
back.  There are better ways to handle this, but I'm leaving it in place
for now.

We don't need explicit calls around the null descriptor write because it
will only be queued if there is room for it, and any scrolling that
might need to be done will _definitely_ be done in the same frame as the
write, since any waiting will have occurred _before_ the descriptor is
written to the buffer.