doc/databuf_cycle_notes: Add notes about CPU cycle timings for databuf loop
These are the sloppy hand calculations that I did to write my
descriptor-complexity-based throttling code for the data buffer.  This
will obviously have to change if the loop is modified, but I wanted to
store this stuff off so that I can sort of remember how I did this in
the first place.

It would be good to automate these calculations at some point, but
that's obviously not trivial, and what I've got now seems to be
working...