kernel/kdata.c: Page-align bufpool
This helps some instructions/addressing modes perform better which (I
think) will be important here soon.

Note that we also zero out the bufpool now to move it into the DATA
section instead of BSS.  This isn't required, but it saves us time
later, because the buffers need to be zeroed before being filled anyway.