filesys: Begin implementing fmount
This is a huge amount of change to pull in in one commit, but I was on a
roll at the time and decided to just go for it.  Realistically I should
have stubbed stuff out and pulled new functions in one-by-one, but I
think this is actually in relatively good shape, despite the insane way
that I tried to pull it all in.

I'm currently hitting a spurious IRQ, but I've already got that figured
out, so I'm going to just keep this as-is.  Fingers crossed...