include/unix.h: Fix TTYDEV index
Now that I've put the "disk" device first in the dev_tab, we need to
shift everything else down by one.  The NDEVS stuff here was already
"correct" (although poorly coded, not using the NDEVS constant), but the
TTYDEV index still being set to 0 was a problem.

This was manifesting in a really interesting way, where I was getting a
spurious IRQ because the code was trying to d_open the TTYDEV as if it
were the disk, and ultimately jumping off into the weeds and executing a
bogus BRK instruction (i.e. uninitialized data), leading to the
unexpected IRQ.

We still hit the "no filesys" panic with this fix, but this is a big
step in the right direction, as I _think_ that's more or less expected
right now.  At very least, it's not totally unexpected.