kernel/head.S: Fix subtle PPU_CTRL initialization bug
After pulling in the huge pile of filesystem/IO related code, PPU
rendering mysteriously broke.  Observation showed that I was correctly
attempting to write the font data to the PPU, but the writes were going
a bit wild on me.

It turns out that totally unrelated code changes had pushed the location
of ppu_ctrl_val out of RAM that was initially zeroed, into RAM that was
set to 0xFF.  Because of this, I was writing 0xFF to PPU_CTRL in my
reset code, which, among other things, caused the PPU to do 32-bit
"vertical" increments during writes, instead of the single-byte
increments I was expecting.

The ppu_ctrl_val will correctly be initialized to zero, as expected, in
the copy_to_ram loop, but this early in boot, we need to explicitly
write 0x00 to PPU_CTRL, in order to achieve the desired behavior.

This was a fun one to debug...