uzi-nes-port: Get filesystem mounting
The filesystem does mount correctly at this point.  However, the code is
a bit ugly, and I'm realizing that having BLKDATA page-aligned isn't
actually going to end up being useful in its current state, as the extra
data off the end of the blkbuf struct ends up causing stuff to be
misaligned anyways, so the bf_data buffers end up starting/ending in
weird spots in pages.  The easiest solution here is probably to just pad
out the blkbuf structs to make them take up some exact number of pages,
but I'm not going to get to that tonight.