Implement wrapped-call PoC
This is just a PoC to make sure that a NOP call wrapper does what's
expected.  The trampoline will eventually be updated to actually page in
the code that we're about to execute, but for now I just want to see
that we can wrap some calls without breaking everything.

Adding call wrappers will make all of the code to up a bit more space,
but it's a necessary evil if we want to be able to page code in and out
at will.