Switch to PRG mode 3
This is pretty much the bare minimum to get PRG mode 3 working
correctly.  There's a lot going on here, but I chose to keep it in one
commit since splitting it up would break bisection to some degree.

The basic pieces that needed to come together here are:

1. Change to PRG mode 3 in head.S
2. Break down the big CODE segment into smaller segments that will fit
   inside of the 8K pages
3. Write a real call-wrapper trampoline to handle paging
4. Figure out some way to handle indirect calls, i.e. various dev_*
   calls in the dev_tab

One extremely subtle, but extremely important piece of getting this to
work was moving the d_open call from filesys.h to devio.h.  This is
something I never would have noticed before, but, since the filesys code
is not paged yet (i.e. still lives in CODE), having d_open defined as if
it were part of it means that d_open will not get paged in when it
should be (i.e. with the rest of the devio.c code).

I had to single step from a known good point before tty_init() in order
to detect this.  It was a pain in the ass.