fceux: Roll back to old uzi branch
The master branch seems to have a bunch of bugs that I wasn't hitting
before.  One was a full lockup with the sound card isn't working right,
and there were also some annoying UI errors where every button/keyboard
shortcut seemed to stop working, but the biggest problem was the loss of
backwards single-stepping in the debugger.

I haven't looked into any of these bugs yet, but I'm here to write NES
code, not debug the emulator (for now at least), so let's roll back to
a point where we know it's working.