unix.h: Generate errno portion of header from Python script
I need to be able to define UZI_* versions of the errno macros in order
to avoid collisions when building with this header included in userspace
Linux tools (i.e. fsutil).

Generate two versions of the macros, one for use inside of UZI, and one
for outside.  This isn't the prettiest solution, but it's better tham
maintaining two versions of the macros.