fsutil: Double back a bit on the design
I discovered the `ucp` tool in Doug's uzi280stuff that handles a lot of
what I'm trying to do here, but maybe a bit smarter.  I want to borrow
from that as much as possible, and avoid duplicating existing code from
the kernel by trying to build/call that stuff directly on the Linux side
of things.