utils/fs/fsutil: Rename fsutil-specific stuff to make the division more clear
Anything built right out of the kernel source should have a linux_
prefix, and anything we build in here to stub things out to make the
Linux code work has an fsutil_ prefix.

This not only makes the naming more meaningful, but also allows us to
have both a linux_filesys.o and an fsutil_filesys.o.