utils/fs/fsutil: Get fsutil's rdtime function working
I had to shuffle stuff around quite a bit to get this to build, because
fsutil.c was trying to pull in UZI's time.h due to the CFLAGS I was
handing it.  This commit ensures that fsutil.c doesn't need to include
_any_ UZI headers, so that we don't get weird issues like that any more.
It also implements what appears to be a functional rdtime.