utils/fs/fsutil: Split out library stuff from header
I don't know if this is totally necessary, but I've already written it,
so I'm doing it.  The problem was that I didn't want to include standard
headers in this file, but I think that might have been a stupid idea
anyway.  Either way, doesn't hurt to have _one_ more file here...