ppu: Implement generic r/w descriptors
This actually doesn't take much.  We basically just need to handle bank
switching, and then the existing write descriptor code can take over.

I did a bit of magic here on the descriptor weight calcultions by just looking
up the cycle times for the added instructions:

LDA (imm)	2 cycles
AND (abs)	4 cycles
BEQ (abs)	4 cycles (max)
JMP (abs)	3 cycles

So that totals 13 additional cycles for r/w descriptor overhead, which I've
bumped to 14 to be safe, making the total overhead for r/w descriptors 115
cycles.  For the per-byte overhead, we just make this the same as the existing
write descriptor overhead, since we have to take the maximum of the two
possibilities here.