kernel: Move udata to the top of kernel RAM
This was a requirement of original UZI that I broke (or just never
followed), but I got away with it until now.  Now that I need to be able
to write blocks to swap during execve, I need to actually be sure that
udata comes first.

This was a bit of a pain to get working, due to some apparent bug/lack
of understanding going on with section alignment stuff.  I've added a
section to the NOTES file along with this commit to describe that
problem.