uzi-nes: Move kernel RAM to highest possible bank
This is to facilitate the changes discussed in previous commits.

As a _huge_ side note, this commit is sort of a poster-child for so many
things are wrong with some of my low-level code right now.  I had to
manually change way too much to much this all work.  It ultimately
didn't take that long to get it figured out, but it really should have
just been some tweaks to a few variables.

For now, I'll leave it as-is, but this needs some serious cleanup.