uzi-nes: Implement doexec
This changes a few conventions here, by necessity.  First off, the
EMAGIC byte is now 0x4c, i.e. an absolute jump.  This is analagous to
what the 0xc3 instruction would be in 8080 Assembly, so I'm assuming
that this is what they intended as well.  This can always change, but it
makes sense for now.

Next, we have to have a new agreed upon way to enter/exit the kernel,
since we don't have the 8080 RST instruction.  This instruction _would_
push the PC onto the stack, and then jump to the address specified, so
we'll do something similar.