include/unix.h: Disambiguate types in shared header
Maybe it's not been the best design choice to try to share the header
with my Linux userspace stuff, but here we are.  I found an issue where
struct cinode was ending up with a different size/offsets in the Linux
fsutil because it was using the builtin int/unsigned ints for stuff.
UZI was expecting 2-byte ints but was seeing 4, and everything came out
garbled.

Here's some output from before I fixed the issue:

<snip>
shell> ls
token ls
argv[0]: ls
fsutil offsetof cinode->c_node: c
kernel sizeof dinode: 40
512    .                01      0  0     20:26:02 03/07/25
512    ..               01      0  0     20:26:02 03/07/25
512    uzi              063     0  0     02:05:52 02/10/25
0      stuff            061     0  0     18:47:42 02/10/25
324    test.c           060     0  0     18:48:06 02/10/25
256    linux_fs.c       057     0  0     19:00:08 02/10/25
16     init             056     0  0     01:11:44 03/07/25
5      dummy            055     0  0     20:24:02 03/07/25
5      dummy2           054     0  0     20:24:28 03/07/25
5      dummy3           053     0  0     20:26:02 03/07/25
</snip>

It's maybe not immediately obvious, but the third column is supposed to
be the mode, so none of that makes sense.  You can also see the c_node
struct coming in at a 12-byte offset instead of 6, which is not right.

Here's the same output after fixing the header:

<snip>
shell> ls
token ls
argv[0]: ls
fsutil offsetof cinode->c_node: 6
512    ./               040777  5  1     20:26:02 03/07/25
512    ../              040777  5  1     20:26:02 03/07/25
512    uzi/             040777  2  51    02:05:52 02/10/25
0      stuff            0100666 1  49    18:47:42 02/10/25
324    test.c           0100666 1  48    18:48:06 02/10/25
256    linux_fs.c       0100666 1  47    19:00:08 02/10/25
16     init             0100666 1  46    01:11:44 03/07/25
5      dummy            0100666 1  45    20:24:02 03/07/25
5      dummy2           0100666 1  44    20:24:28 03/07/25
5      dummy3           0100666 1  43    20:26:02 03/07/25
</snip>

I'm actually surprised that everything comes out nice now, without
having to re-roll the filesystem.  It might still be corrupt in some
other way, but that's easy to deal with.

I _think_ since unix.h is the only shared header, this should be the end
of this, but there could easily be more problems like this lurking
around...