uzi-nes: Properly move PROGBASE up to 0x6000
This won't currently work, or will at very least clobbber the kernel
stacks, but it's a first step here.  I need to do some work to page out
the kernel stack before copying the program in in order to avoid this.

Honestly, I'm not sure what I was thinking with the original 0xD000
definition here...