uzi-nes: Shuffle banks around again...
I feel like this is a setback, but for now it's just how it has to be
done.

I tried really hard to get a "single bank" setup implemented for the
kernel, which almost sort of makes sense, if you bank for every call,
but it falls apart because you really _need_ some stuff to be in memory
all the time.  Most specifically, all the primitive functions emitted by
the C compiler that are needed for basic C code to function.  There's
also the trampoline code, which was do-able by adding it to an end stub
for each segment, but wasn't ideal and, obviously, it all falls apart
when you page out the C compiler primitives...

So, the solution for now just has to be to only have 16K available for
user processes.  This sucks, but I think I _can_ make it better, it's
just going to take some work, and I'd really rather get some basic
functionality in place before I start trying to get super-crazy with new
banking strategies.

Note the very important change here to _not_ bank out kernel RAM during
tempstack.  This caused quite a headache to debug, because things
obviously behave a bit weird if you rip the memory out from under the
code.

Anyway, this gets us to a state where we panic on a corrupt inode while
trying to load in the user process.  It's possible that my port of the
ucp tool is corrupting the filesystem, or possible that there's just a
bug here, but that's up next.