uzi-nes: Fix banking for userspace switch
There were several things wrong here:

  1. I was swapping the wrong bank during interrupts.

This is an easy fix - just use the right bank.

  2. The first user bank needs to be paged in _before_ we start reading
     in the program from disk.

I created the user_bank_setup function to handle this.  It does the
first page swap and tempstack handles the second page (i.e. after it's
actually safe to get rid of the kernel stack).

  3. I wasn't actually setting the stack pointer where I thought I was.

The use of immediate vs indirect addressing with the setting of the
temp_stack pointer was giving interesting results.  This one was hard to
find, but easy to understand once I spotted it.

  4. jump_test was off by one byte.

This was another hard one to spot.  An RTS would have added a byte to
to the value pulled from the stack before returning to execution, so we
need to do the same, in order to avoid landing at a bogus address.