kernel/scall2.c: Fix post-program-load bzero
Since progptr is pointing to the end of the first block of the program
(i.e. 0x6200), I'm pretty sure this was causing us to do a bzero with
size -0x200.  The size on bzero is unsigned, so I'm sure it was just
interpreting this as some weirdly large value (0xfe00 I think?).

Anyway, this is obviously wrong, so fix it.