uzi-nes: Fix ppuwd writes
This had never really been properly tested, so not super surprising that
it didn't work...

The first problem here, which isn't directly address, but is important
to mention, is that a write descriptor for 64 bytes is slightly too
large for our current PPU_BUF_SIZE.  This meant that the descriptor was
getting rejected in an infinite loop, waiting for there to be room in
the buffer, when this was never going to happen.

Realistically, we can make room for these descriptors, as a 64 byte
write should totally fit into one VBlank (I think it would be ~1800
cycles, from observation), but I didn't want to rework that stuff right
now.

Next there was a one-off issue with the PPU bank descriptor low byte,
and the data copy into the actual PPU buffer.  This is the FIXME bit in
__queue_descriptor, where we add a "magic 1" to the descriptor size, to
make room for the low bank byte, which is _not_ considered part of the
data by process_ppubuf.