ppu/asm/ppu_databuf_asm.S: Fix subtle, gross error in process_ppubuf
This code was written a _long_ time ago, and I tried to make it
general-use for ppuwd writes, as well as character writes to video
memory.  This doesn't _quite_ work though, because the code was never
actually able to process multi-byte write descriptors, video or
otherwise.

When a multi-byte descriptor was supplied, we actually ended up just
writing the first byte over and over.  We wrote the correct number of
bytes, just not the correct bytes...  Anyway this is dead simple to fix,
by just actually incrementing Y while we loop through the data.

Note that this exposed some other flaws/shortcomings of process_ppubuf,
which I'll address subsequent commits.